Move impl to dummy/
